requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
psycopg[binary]
waitress==3.0.0
python-dotenv==1.0.1
alembic==1.13.2